#8428: Migrate the sweep tests to sweep framework
